+++ R/dof.R
@@ -188,7 +188,7 @@ dof <- degrees_of_freedom
 
 #' @keywords internal
 .degrees_of_freedom_residual <- function(model, verbose = TRUE) {
-  if (.is_bayesian_model(model) && !inherits(model, c("bayesx", "blmerMod", "bglmerMod"))) {
+  if (.is_bayesian_model(model, exclude = c("bmerMod", "bayesx", "blmerMod", "bglmerMod"))) {
     model <- bayestestR::bayesian_as_frequentist(model)
   }
 
@@ -277,7 +277,7 @@ dof <- degrees_of_freedom
       return(TRUE)
     } else {
       if (verbose) {
-        insight::format_alert(sprintf("`%s` must be one of \"wald\", \"residual\" or \"profile\". Using \"wald\" now.", type))
+        insight::format_alert(sprintf("`%s` must be one of \"wald\", \"residual\" or \"profile\". Using \"wald\" now.", type)) # nolint
       }
       return(FALSE)
     }
@@ -289,7 +289,7 @@ dof <- degrees_of_freedom
       return(TRUE)
     } else {
       if (verbose) {
-        insight::format_alert(sprintf("`%s` must be one of \"wald\", \"normal\" or \"boot\". Using \"wald\" now.", type))
+        insight::format_alert(sprintf("`%s` must be one of \"wald\", \"normal\" or \"boot\". Using \"wald\" now.", type)) # nolint
       }
       return(FALSE)
     }
@@ -298,24 +298,24 @@ dof <- degrees_of_freedom
   info <- insight::model_info(model, verbose = FALSE)
   if (!is.null(info) && isFALSE(info$is_mixed) && method == "boot") {
     if (verbose) {
-      insight::format_alert(sprintf("`%s=boot` only works for mixed models of class `merMod`. To bootstrap this model, use `bootstrap=TRUE, ci_method=\"bcai\"`.", type))
+      insight::format_alert(sprintf("`%s=boot` only works for mixed models of class `merMod`. To bootstrap this model, use `bootstrap=TRUE, ci_method=\"bcai\"`.", type)) # nolint
     }
     return(TRUE)
   }
 
   if (is.null(info) || !info$is_mixed) {
-    if (!(method %in% c("analytical", "any", "fit", "betwithin", "nokr", "wald", "ml1", "profile", "boot", "uniroot", "residual", "normal"))) {
+    if (!(method %in% c("analytical", "any", "fit", "betwithin", "nokr", "wald", "ml1", "profile", "boot", "uniroot", "residual", "normal"))) { # nolint
       if (verbose) {
-        insight::format_alert(sprintf("`%s` must be one of \"residual\", \"wald\", \"normal\", \"profile\", \"boot\", \"uniroot\", \"betwithin\" or \"ml1\". Using \"wald\" now.", type))
+        insight::format_alert(sprintf("`%s` must be one of \"residual\", \"wald\", \"normal\", \"profile\", \"boot\", \"uniroot\", \"betwithin\" or \"ml1\". Using \"wald\" now.", type)) # nolint
       }
       return(FALSE)
     }
     return(TRUE)
   }
 
-  if (!(method %in% c("analytical", "any", "fit", "satterthwaite", "betwithin", "kenward", "kr", "nokr", "wald", "ml1", "profile", "boot", "uniroot", "residual", "normal"))) {
+  if (!(method %in% c("analytical", "any", "fit", "satterthwaite", "betwithin", "kenward", "kr", "nokr", "wald", "ml1", "profile", "boot", "uniroot", "residual", "normal"))) { # nolint
     if (verbose) {
-      insight::format_alert(sprintf("`%s` must be one of \"residual\", \"wald\", \"normal\", \"profile\", \"boot\", \"uniroot\", \"kenward\", \"satterthwaite\", \"betwithin\" or \"ml1\". Using \"wald\" now.", type))
+      insight::format_alert(sprintf("`%s` must be one of \"residual\", \"wald\", \"normal\", \"profile\", \"boot\", \"uniroot\", \"kenward\", \"satterthwaite\", \"betwithin\" or \"ml1\". Using \"wald\" now.", type)) # nolint
     }
     return(FALSE)
   }
@@ -335,12 +335,17 @@ dof <- degrees_of_freedom
 
 # helper
 
-.is_bayesian_model <- function(x) {
-  inherits(x, c(
+.is_bayesian_model <- function(x, exclude = NULL) {
+  bayes_classes <- c(
     "brmsfit", "stanfit", "MCMCglmm", "stanreg",
     "stanmvreg", "bmerMod", "BFBayesFactor", "bamlss",
     "bayesx", "mcmc", "bcplm", "bayesQR", "BGGM",
     "meta_random", "meta_fixed", "meta_bma", "blavaan",
-    "blrm"
-  ))
+    "blrm", "blmerMod"
+  )
+  # if exclude is not NULL, remove elements in exclude from bayes_class
+  if (!is.null(exclude)) {
+    bayes_classes <- bayes_classes[!bayes_classes %in% exclude]
+  }
+  inherits(x, bayes_classes)
 }